完成全部转为列存
comment 		264.4
date 			36.1
discount 		        24
extend_price	        24
line_num		        6
linestatus 		6
orderkey 		6.2
parkey 			24
quantity 		        6
return_flag 	        6
ship 			6
supperkey 		24
tax 			        24

+++ src/observer/defs.h
@@ -15,6 +15,6 @@ See the Mulan PSL v2 for more details. */
 #pragma once
 
 using PageNum = int16_t;
-using SlotNum = int16_t;
+using SlotNum = int32_t;
 
 const uint32_t MAX_LINE_NUM = 6010000;
